Push some structs down into their own modules
