+++ src/lib.rs
@@ -1,17 +1,20 @@
-use regex::Regex;
 use serde::Serialize;
+use std::collections::HashMap;
 #[cfg(test)]
 use std::ptr;
-use std::{
-    collections::HashMap,
-    ffi::OsStr,
-    fmt,
-    fs::{self, File},
-    io,
-    ops::Range,
-    path::PathBuf,
-};
-use tree_sitter::{Language, Node, Parser, Query, QueryCursor, Range as TSRange, Tree};
+
+mod call_graph;
+mod code_source;
+mod source_query;
+mod source_ref;
+
+// TODO: doesn't need to be exposed if we can clean up the arguments to do_mapping
+pub use call_graph::CallGraph;
+use call_graph::Edge;
+pub use code_source::CodeSource;
+use source_query::QueryResult;
+pub use source_query::SourceQuery;
+pub use source_ref::SourceRef;
 
 pub struct Filter {
     pub start: usize,
@@ -76,76 +79,6 @@ impl SourceLanguage {
     }
 }
 
-pub struct CodeSource {
-    filename: String,
-    language: SourceLanguage,
-    buffer: String,
-}
-
-const SUPPORTED_EXTS: &[&str] = &["java", "rs"];
-
-impl CodeSource {
-    fn new(path: PathBuf, mut input: Box<dyn io::Read>) -> CodeSource {
-        let language = match path.extension() {
-            Some(ext) => match ext.to_str().unwrap() {
-                "rs" => SourceLanguage::Rust,
-                "java" => SourceLanguage::Java,
-                _ => panic!("Unsupported language"),
-            },
-            None => panic!("No extension"),
-        };
-        let mut buffer = String::new();
-        input.read_to_string(&mut buffer).expect("can read source");
-        CodeSource {
-            language,
-            filename: path.to_string_lossy().to_string(),
-            buffer,
-        }
-    }
-
-    fn ts_language(&self) -> Language {
-        match self.language {
-            SourceLanguage::Rust => tree_sitter_rust::language(),
-            SourceLanguage::Java => tree_sitter_java::language(),
-        }
-    }
-}
-
-pub fn find_code(sources: &str) -> Vec<CodeSource> {
-    let mut srcs = vec![];
-    let meta = fs::metadata(sources).expect("can read file metadata");
-    if meta.is_file() {
-        let path = PathBuf::from(sources);
-        try_add_file(path, &mut srcs);
-    } else {
-        walk_dir(PathBuf::from(sources), &mut srcs).expect("can traverse directory");
-    }
-    srcs
-}
-
-fn walk_dir(dir: PathBuf, srcs: &mut Vec<CodeSource>) -> io::Result<()> {
-    for entry in fs::read_dir(dir)? {
-        let entry = entry?;
-        let path = entry.path();
-        let metadata = fs::metadata(&path)?;
-        if metadata.is_file() {
-            try_add_file(path, srcs);
-        } else if metadata.is_dir() {
-            walk_dir(path, srcs).expect("can traverse directory");
-        }
-    }
-    Ok(())
-}
-
-fn try_add_file(path: PathBuf, srcs: &mut Vec<CodeSource>) {
-    let ext = path.extension().unwrap_or(OsStr::new(""));
-    if SUPPORTED_EXTS.iter().any(|&supported| supported == ext) {
-        let input = Box::new(File::open(PathBuf::from(&path)).expect("can open file"));
-        let code = CodeSource::new(path, input);
-        srcs.push(code);
-    }
-}
-
 #[derive(Serialize)]
 pub struct LogMapping<'a> {
     #[serde(skip_serializing)]
@@ -161,170 +94,10 @@ pub struct LogRef<'a> {
     pub line: &'a str,
 }
 
-pub struct QueryResult {
-    kind: String,
-    range: TSRange,
-    name_range: Range<usize>,
-}
-
-pub struct SourceQuery<'a> {
-    pub source: &'a str,
-    tree: Tree,
-    language: Language,
-}
-
-impl<'a> SourceQuery<'a> {
-    pub fn new(code: &'a CodeSource) -> SourceQuery<'a> {
-        // println!("{}", code.filename);
-        let mut parser = Parser::new();
-        let language = code.ts_language();
-        parser
-            .set_language(&language)
-            .expect(format!("Error loading {:?} grammar", language).as_str());
-        let source = code.buffer.as_str();
-        let tree = parser.parse(source, None).expect("source is parsable");
-        // println!("{:?}", tree.root_node().to_sexp());
-        SourceQuery {
-            source,
-            tree,
-            language,
-        }
-    }
-
-    pub fn query(&self, query: &str, node_kind: Option<&str>) -> Vec<QueryResult> {
-        let query = Query::new(&self.language, query).unwrap();
-        let filter_idx = node_kind.map_or(None, |kind| query.capture_index_for_name(kind));
-        let mut cursor = QueryCursor::new();
-        cursor
-            .matches(&query, self.tree.root_node(), self.source.as_bytes())
-            .into_iter()
-            .flat_map(|m| m.captures)
-            .filter(|c| {
-                filter_idx.is_none() || (filter_idx.is_some() && filter_idx.unwrap() == c.index)
-            })
-            .map(|c| QueryResult {
-                kind: String::from(c.node.kind()),
-                range: c.node.range(),
-                name_range: self.find_fn_range(c.node),
-            })
-            .collect()
-    }
-
-    fn find_fn_range(&self, node: Node) -> Range<usize> {
-        // println!("node.kind()={:?}", node.kind());
-        match node.kind() {
-            "function_item" => {
-                let range = node.child_by_field_name("name").unwrap().range();
-                range.start_byte..range.end_byte
-            }
-            "method_declaration" => {
-                let range = node.child_by_field_name("name").unwrap().range();
-                range.start_byte..range.end_byte
-            }
-            "constructor_declaration" => {
-                let range = node.child_by_field_name("name").unwrap().range();
-                range.start_byte..range.end_byte
-            }
-            "class_declaration" => {
-                let range = node.child_by_field_name("name").unwrap().range();
-                range.start_byte..range.end_byte
-            }
-            _ => {
-                let r = self.find_fn_range(node.parent().unwrap());
-                // println!("*****");
-                r
-            }
-        }
-    }
-}
-
-// TODO: get rid of this clone?
-#[derive(Clone, Debug, Serialize)]
-pub struct SourceRef {
-    #[serde(rename(serialize = "sourcePath"))]
-    source_path: String,
-    #[serde(rename(serialize = "lineNumber"))]
-    pub line_no: usize,
-    column: usize,
-    name: String,
-    text: String,
-    #[serde(skip_serializing)]
-    matcher: Regex,
-    vars: Vec<String>,
-}
-
-impl fmt::Display for SourceRef {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "[Line: {}, Col: {}] source `{}` name `{}` vars={:?}",
-            self.line_no, self.column, self.text, self.name, self.vars
-        )
-    }
-}
-
-impl PartialEq for SourceRef {
-    fn eq(&self, other: &Self) -> bool {
-        self.line_no == other.line_no
-            && self.column == other.column
-            && self.name == other.name
-            && self.text == other.text
-            && self.vars == other.vars
-    }
-}
-
-#[derive(Debug)]
-pub struct CallGraph<'a> {
-    edges: Vec<Edge<'a>>,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct Edge<'a> {
-    // same as SourceRef found in via
-    // from: &'a str,
-    to: &'a str,
-    via: SourceRef,
-}
-
-impl<'a> CallGraph<'a> {
-    pub fn new(sources: &'a mut Vec<CodeSource>) -> CallGraph<'a> {
-        let edges = Self::find_edges(sources);
-        CallGraph { edges }
-    }
-
-    fn find_edges(sources: &'a mut Vec<CodeSource>) -> Vec<Edge<'a>> {
-        let mut symbols = Vec::new();
-        let edge_query = r#"
-            (call_expression function: (identifier) @fn_name arguments: (arguments (_))*)
-        "#;
-        for code in sources.iter() {
-            if code.language == SourceLanguage::Rust {
-                let src_query = SourceQuery::new(code);
-                let results = src_query.query(edge_query, Some("fn_name"));
-
-                for result in results {
-                    let range = result.range;
-                    let fn_call = &src_query.source[range.start_byte..range.end_byte];
-                    let src_ref = build_src_ref(code, result);
-
-                    symbols.push(Edge {
-                        to: fn_call,
-                        via: src_ref,
-                    });
-                }
-            }
-        }
-        symbols
-    }
-}
-
 pub fn link_to_source<'a>(log_ref: &LogRef, src_refs: &'a [SourceRef]) -> Option<&'a SourceRef> {
-    src_refs.iter().find(|&source_ref| {
-        if let Some(_) = source_ref.matcher.captures(log_ref.line) {
-            return true;
-        }
-        false
-    })
+    src_refs
+        .iter()
+        .find(|&source_ref| source_ref.captures(log_ref).is_some())
 }
 
 pub fn extract_variables<'a>(
@@ -333,7 +106,7 @@ pub fn extract_variables<'a>(
 ) -> HashMap<&'a str, &'a str> {
     let mut variables = HashMap::new();
     if src_ref.vars.len() > 0 {
-        if let Some(captures) = src_ref.matcher.captures(log_line.line) {
+        if let Some(captures) = src_ref.captures(&log_line) {
             for i in 0..captures.len() - 1 {
                 variables.insert(
                     src_ref.vars[i].as_str(),
@@ -439,7 +212,7 @@ pub fn extract_logging<'a>(sources: &mut Vec<CodeSource>) -> Vec<SourceRef> {
             // println!("node.kind()={:?} range={:?}", result.kind, result.range);
             match result.kind.as_str() {
                 "string_literal" => {
-                    let src_ref = build_src_ref(code, result);
+                    let src_ref = SourceRef::new(code, result);
                     matched.push(src_ref);
                 }
                 "identifier" | "this" => {
@@ -465,50 +238,6 @@ pub fn extract_logging<'a>(sources: &mut Vec<CodeSource>) -> Vec<SourceRef> {
         }
     }
     matched
-}
-
-fn build_src_ref<'a, 'q>(code: &CodeSource, result: QueryResult) -> SourceRef {
-    let range = result.range;
-    let source = code.buffer.as_str();
-    let text = source[range.start_byte..range.end_byte].to_string();
-    let line = range.start_point.row + 1;
-    let col = range.start_point.column;
-    let start = range.start_byte + 1;
-    let mut end = range.end_byte - 1;
-    if start == range.end_byte {
-        end = range.end_byte;
-    }
-    let unquoted = &source[start..end].to_string();
-    // println!("{} line {}", code.filename, line);
-    let matcher = build_matcher(unquoted);
-    let vars = Vec::new();
-    let name = source[result.name_range].to_string();
-    SourceRef {
-        source_path: code.filename.clone(),
-        line_no: line,
-        column: col,
-        name,
-        text,
-        matcher,
-        vars,
-    }
-}
-
-fn build_matcher(text: &str) -> Regex {
-    // XXX: avoid regex that are too greedy by returning a regex that
-    //      never matches anything
-    if text == "{}" || text.trim() == "" {
-        Regex::new(r#"\w\b\w"#).unwrap()
-    } else {
-        let curly_replacer = Regex::new(r#"\\?\{.*?\}"#).unwrap();
-        let escaped = curly_replacer
-            .split(text)
-            .map(|s| regex::escape(s))
-            .collect::<Vec<String>>()
-            .join(r#"(\w+)"#);
-        // println!("escaped = {}", Regex::new(&escaped).unwrap().as_str());
-        Regex::new(&escaped).unwrap()
-    }
 }
 
 #[test]
@@ -681,22 +410,4 @@ fn test_find_possible_paths() {
         vars: vec![],
     };
     assert_eq!(paths, vec![vec![&foo_2_nope, &main_2_foo]])
-}
-
-#[test]
-fn test_build_matcher_needs_escape() {
-    let matcher = build_matcher("{}) {}, {}");
-    assert_eq!(
-        Regex::new(r#"(\w+)\) (\w+), (\w+)"#).unwrap().as_str(),
-        matcher.as_str()
-    );
-}
-
-#[test]
-fn test_build_matcher_mix() {
-    let matcher = build_matcher("{}) {:?}, {foo.bar}");
-    assert_eq!(
-        Regex::new(r#"(\w+)\) (\w+), (\w+)"#).unwrap().as_str(),
-        matcher.as_str()
-    );
 }
